+++ blender/source/blender/draw/intern/draw_view.cc
@@ -1,5 +1,6 @@
-/* SPDX-License-Identifier: GPL-2.0-or-later
- * Copyright 2022 Blender Foundation. */
+/* SPDX-FileCopyrightText: 2022 Blender Authors
+ *
+ * SPDX-License-Identifier: GPL-2.0-or-later */
 
 /** \file
  * \ingroup draw
@@ -44,14 +45,14 @@ void View::frustum_boundbox_calc(int view_id)
 {
   /* Extract the 8 corners from a Projection Matrix. */
 #if 0 /* Equivalent to this but it has accuracy problems. */
-  BKE_boundbox_init_from_minmax(&bbox, float3(-1.0f),float3(1.0f));
+  BKE_boundbox_init_from_minmax(&bbox, float3(-1.0f), float3(1.0f));
   for (int i = 0; i < 8; i++) {
     mul_project_m4_v3(data_.wininv.ptr(), bbox.vec[i]);
   }
 #endif
 
   MutableSpan<float4> corners = {culling_[view_id].frustum_corners.corners,
-                                 ARRAY_SIZE(culling_[view_id].frustum_corners.corners)};
+                                 int64_t(ARRAY_SIZE(culling_[view_id].frustum_corners.corners))};
 
   float left, right, bottom, top, near, far;
   bool is_persp = data_[view_id].winmat[3][3] == 0.0f;
@@ -107,7 +108,7 @@ void View::frustum_culling_sphere_calc(int view_id)
 {
   BoundSphere &bsphere = *reinterpret_cast<BoundSphere *>(&culling_[view_id].bound_sphere);
   Span<float4> corners = {culling_[view_id].frustum_corners.corners,
-                          ARRAY_SIZE(culling_[view_id].frustum_corners.corners)};
+                          int64_t(ARRAY_SIZE(culling_[view_id].frustum_corners.corners))};
 
   /* Extract Bounding Sphere */
   if (data_[view_id].winmat[3][3] != 0.0f) {
